add witches curse to shoot
